updates, move Loader.prototype.parse into loader implementation
