+++ package.json
@@ -41,7 +41,7 @@
     "test": "cd test && node test"
   },
   "dependencies": {
-    "traceur": "0.0.49",
-    "when": "^3.2.3"
+    "traceur": "0.0.52",
+    "when": "^3.4.2"
   }
 }
